Split sidebar render item id
